With global adoption of even/odd decimators the accu type must be 32 not 64 bits

+++ plugins/samplesource/airspy/airspythread.h
@@ -56,7 +56,7 @@ private:
 	static AirspyThread *m_this;
 
 #ifdef SDR_RX_SAMPLE_24BIT
-    Decimators<qint64, qint16, SDR_RX_SAMP_SZ, 12> m_decimators;
+    Decimators<qint32, qint16, SDR_RX_SAMP_SZ, 12> m_decimators;
 #else
 	Decimators<qint32, qint16, SDR_RX_SAMP_SZ, 12> m_decimators;
 #endif
