+++ plugins/samplesource/hackrfinput/hackrfinputthread.h
@@ -55,7 +55,7 @@ private:
 	int m_fcPos;
 
 #ifdef SDR_RX_SAMPLE_24BIT
-    Decimators<qint64, qint8, SDR_RX_SAMP_SZ, 8> m_decimators;
+    Decimators<qint32, qint8, SDR_RX_SAMP_SZ, 8> m_decimators;
 #else
 	Decimators<qint32, qint8, SDR_RX_SAMP_SZ, 8> m_decimators;
 #endif
