+++ plugins/samplesource/plutosdrinput/plutosdrinputthread.h
@@ -60,7 +60,7 @@ private:
     float m_phasor;
 
 #ifdef SDR_RX_SAMPLE_24BIT
-    Decimators<qint64, qint16, SDR_RX_SAMP_SZ, 12> m_decimators;
+    Decimators<qint32, qint16, SDR_RX_SAMP_SZ, 12> m_decimators;
 #else
     Decimators<qint32, qint16, SDR_RX_SAMP_SZ, 12> m_decimators;
 #endif
